requirements.txt
# API and Web Framework
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7

# Database and Caching
redis>=4.5.4
psycopg2-binary>=2.9.6
sqlalchemy>=2.0.9
sqlalchemy-utils>=0.41.1
alembic>=1.10.3

# Data Fetching and Processing
yfinance>=0.2.18
pandas>=2.0.0
numpy>=1.24.2
scrapy>=2.8.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.24.0
pyyaml>=6.0